% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acei_arb_meds.R
\name{acei_arb_meds}
\alias{acei_arb_meds}
\title{ACEI/ARB Meds}
\usage{
acei_arb_meds(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
CSV saved to output_folder/acei_arb_meds.csv
}
\description{
ACEI/ARB Meds
}
\details{
Meds: "fosinopril", "fosinopril sodium","monopril", "ramipril", "altace", "captopril", "capoten",
"moexipril", "univasc", "lisinopril", "zestril", "prinivil", "enalapril", "vasotec", "epaned", "quinapril",
"accupril", "trandolapril", "mavik", "gopten", "odrik", "benazepril", "lotensin", "perindopril", "aceon"
"eprosartan", "teveten", "azilsartan", "medoxomil", "edarbi", "olmesartan", "benicar", "valsartan", "diovan",
"telmisartan", "micardis", "losartan",
"cozaar", "candesartan", "atacand", "irbesartan", "avapro"
}
