% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acute_heart_failure.R
\name{acute_heart_failure}
\alias{acute_heart_failure}
\title{Acute Heart Failure}
\usage{
acute_heart_failure(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/acute_heart_failure.csv
}
\description{
Acute Heart Failure
}
\details{
At least 1 ICD code:

ICD9: 428.21, 429.23, 428.31, 428.33, 428.41, 428.43

ICD10: I50.21, I50.23, I50.31, I50.33, I50.41, I50.43
}
