% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afib.R
\name{afib}
\alias{afib}
\title{Afib}
\usage{
afib(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL,
  suffix = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}

\item{suffix}{optional string appended to the end of every output column name except person_id.}
}
\value{
output_folder/afib.csv
}
\description{
Afib
}
\details{
At least 1 ICD code:

ICD9: "427.3","427.31","427.32"

ICD10: "I48","I48.0","I48.1","I48.2","I48.3","I48.4","I48.9","I48.91","I48.92"
}
