% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_today.R
\name{age_today}
\alias{age_today}
\title{Age Today}
\usage{
age_today(output_folder)
}
\arguments{
\item{output_folder}{the folder to write the output}
}
\value{
output_folder/age_today.csv
}
\description{
Age Today
}
