% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alcohol_drinks_typical.R
\name{alcohol_drinks_typical}
\alias{alcohol_drinks_typical}
\title{Number of drinks on a typical day}
\usage{
alcohol_drinks_typical(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/alcohol_drinks_typical.csv
}
\description{
Number of drinks on a typical day
}
\details{
Alcohol AoU intake survey. On a typical day when you drink, how many drinks do you have?
}
