% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_alt.R
\name{all_alt}
\alias{all_alt}
\title{All ALT}
\usage{
all_alt(output_folder, anchor_date_table = NULL, before = NULL, after = NULL)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/all_alt.csv
}
\description{
All ALT
}
\details{
Searches for

"Alanine aminotransferase"
"Alanine aminotransferase | Serum or Plasma | Chemistry - non-challenge"
"Alanine aminotransferase \[Enzymatic activity/volume] in Serum or Plasma"
}
