% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_cabg_codes.R
\name{all_cabg_codes}
\alias{all_cabg_codes}
\title{All cabg codes}
\usage{
all_cabg_codes(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/all_cabg_codes.csv
CPT: "33510","33511","33512","33513","33514","33516","35600","33533","33534","33535","33536","1006199","1006200","1006208","1006216","1006217","4110F","33523","33522","33521","33519","33518","33517"
}
\description{
All cabg codes
}
