% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_co2.R
\name{all_carbon_dioxide}
\alias{all_carbon_dioxide}
\title{All CO2}
\usage{
all_carbon_dioxide(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/all_carbon_dioxide.csv
}
\description{
All CO2
}
\details{
Searches for

"Carbon dioxide"
"Carbon dioxide | Serum or Plasma | Chemistry - non-challenge"
"Carbon dioxide, total \link{Moles/volume} in Serum or Plasma"
}
