% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_dcm_codes.R
\name{all_dcm_codes}
\alias{all_dcm_codes}
\title{All DCM codes}
\usage{
all_dcm_codes(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/all_dcm_codes.csv
ICD9: "425.5","674.50","674.51","674.52","674.53","674.54"

ICD10: "I42.0","A36.81","B33.24","I42.6","I42.7","O90.3"
}
\description{
All DCM codes
}
