% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_hgba1c.R
\name{all_hgba1c}
\alias{all_hgba1c}
\title{All HgbA1c}
\usage{
all_hgba1c(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/all_hgba1c.csv
}
\description{
All HgbA1c
}
\details{
Searches for

"Hemoglobin A1c/Hemoglobin.total in Blood by Electrophoresis"

"Hemoglobin A1c/Hemoglobin.total in Blood by calculation"

"Hemoglobin A1c/Hemoglobin.total in Blood by IFCC protocol"

"Hemoglobin A1c/Hemoglobin.total in Blood"

"Hemoglobin A1c/Hemoglobin.total in Blood by HPLC"
}
