% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_ldl.R
\name{all_ldl}
\alias{all_ldl}
\title{Closest LDL}
\usage{
all_ldl(output_folder, anchor_date_table = NULL, before = NULL, after = NULL)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/all_ldl.csv
}
\description{
Closest LDL
}
\details{
Searches for

"Cholesterol in LDL [Mass/volume] in Serum or Plasma by calculation"

"Cholesterol in LDL [Mass/volume] in Serum or Plasma"

"Cholesterol in LDL [Mass/volume] in Serum or Plasma by Direct assay"

"Cholesterol in LDL [Mass/volume] in Serum or Plasma by Electrophoresis"
}
