% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_mi_codes.R
\name{all_mi_codes}
\alias{all_mi_codes}
\title{All MI codes}
\usage{
all_mi_codes(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/all_mi_codes.csv
ICD9: "410.00","410.01","410.02","410.10","410.11","410.12","410.20","410.21","410.22","410.30","410.31","410.32","410.40","410.41","410.42","410.50","410.51","410.52","410.80","410.81","410.82","410.90","410.91","410.92","411.0","412","429.79"

ICD10: "I21", "I21.0","I21.01","I21.02","I21.09","I21.1","I21.11","I21.19","I21.2","I21.21","I21.29","I21.3","I21.4","I21.9","I21.A","I21.A1","I21.A9","I21.B","I22","I22.0","I22.1","I22.2","I22.8","I22.9","I23","I23.0","I23.1","I23.2","I23.3","I23.4","I23.5","I23.6","I23.8","I24.1","I25.2"
}
\description{
All MI codes
}
