% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_pci_codes.R
\name{all_pci_codes}
\alias{all_pci_codes}
\title{All pci codes}
\usage{
all_pci_codes(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/all_pci_codes.csv
CPT: "92944","92943","92941","92938","92937","92934","92933","92929","92928","92924","92925","92921","92920","37247",
"37246","1029697","1021168","1021167","1021166","1021165","1021164","1021163","0715T","0659T"
}
\description{
All pci codes
}
