% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacterial_infection.R
\name{bacterial_infection}
\alias{bacterial_infection}
\title{Bacterial infection}
\usage{
bacterial_infection(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/bacterial_infection.csv
}
\description{
Bacterial infection
}
\details{
At least 1 ICD 10 code (ICD 9 is not included):

Pneumonia/respiratory
J18.9: Unspecified Pneumonia
J15.0: Pneumonia due to Klebsiella pneumoniae
J15.1: Pneumonia due to Pseudomonas
J15.2: Pneumonia due to Staphylococcus
J15.3: Pneumonia due to Streptococcus group B
J15.4: Pneumonia due to other streptococci
J15.6: Pneumonia due to other Gram-negative bacteria
J15.7: Pneumonia due to Mycoplasma pneumoniae
J15.8: Pneumonia due to other specified bacteria
J18.0: Bronchopneumonia, unspecified organism
J18.8: Other pneumonia, unspecified organism
J86.0: Empyema
J85: Lung abscess
J01.0: Acute maxillary sinusitis
J01.1: Acute frontal sinusitis
J01.2: Acute ethmoidal sinusitis
J01.3: Acute sphenoidal sinusitis
J01.4: Acute pansinusitis
J01.8: Other acute sinusitis
J01.9: Acute sinusitis, unspecified

Genitourinary infections
N39.0: Urinary tract infection, site not specified
N39.1: Cystitis (bladder infection)
N39.2: Pyelonephritis (kidney infection)
A54.0: Gonorrhea
A56.0: Chlamydia
A54.29: Other gonococcal genitourinary infections
A56.19: Other chlamydial genitourinary infections

Skin and soft tissue infections
L00-L02: Abscesses and furuncles
L03: Cellulitis and lymphangitis
L06: Necrosis of skin and subcutaneous tissue
L07: Myositis
L08: Other infections of skin and subcutaneous tissue, unspecified

Bone and joint infections
M86.x: Osteomyelitis
M00.x: Septic Artiritis
T84.50: infection and inflammatory reaction due to an unspecified internal joint prosthesis.
T84.7: infection due to other internal orthopedic prosthetic devices, implants, and grafts.
M46.24: Osteomyelitis of vertebra, thoracic region
M46.34: Infection of intervertebral disc (pyogenic), thoracic region:

Gastrointestinal infections
K52.9: Unspecified infectious gastroenteritis and colitis
A00-A01: Cholera
A02: Salmonella infections
A03: Shigellosis
A04: Other bacterial intestinal infections
A04.9: Bacterial intestinal infection, unspecified
K81.0: Acute cholecystitis
K81.2: Acute cholecystitis with chronic cholecystitis
K81.9: Cholecystitis, unspecified
CNS infections
G00.0: Hemophilus influenzae meningitis
G00.1: Pneumococcal meningitis
G00.2: Streptococcal meningitis
G00.3: Staphylococcal meningitis
G00.8: Other bacterial meningitis
G00.9: Bacterial meningitis, unspecified
A39: Meningococcal infections
A40: Other bacterial meningitis
}
