% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_blockers.R
\name{beta_blockers}
\alias{beta_blockers}
\title{Beta blockers}
\usage{
beta_blockers(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/beta_blockers.csv
}
\description{
Beta blockers
}
\details{
Meds: "betaxolol", "kerlone", "betoptic", "acebutolol", "sectral",
"atenolol", "tenormin", "metoprolol", "metoprolol succinate",
"toprol-xl", "metoprolol tartrate", "lopressor", "metoprolol succinate er",
"metoprolol tartrate", "dutoprolol", "nebivolol", "bystolic", "bisoprolol", "zebeta", "esmolol",
"brevibloc", "propranolol", "innopran xl", "inderal", "inderal la", "nadolol",
"corgard", "carvedilol", "coreg", "coreg cr", "labetalol", "normodyne",
"trandate", "timolol", "blocadren", "timoptic", "istalol", "betimol", "carteolol",
"cartrol", "penbutolol", "levatol", "pindolol", "visken"
}
