% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bout_cadence.R
\name{bout_cadence}
\alias{bout_cadence}
\title{Bout Cadence}
\usage{
bout_cadence(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL,
  min_steps = 60
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}

\item{min_steps}{a numeric value indicating the minimum number of steps per minute to be included in a bout. Default is 60.}
}
\value{
output_folder/bout_cadence.csv
}
\description{
Bout Cadence
}
\details{
Bout cadence is the average steps per minute when the step count >/= \code{min_steps} steps a minute for at least 2 minutes.
The wearer will usually have many of these “bouts” throughout the day.
We take the average over the entire day to get the average bout cadence.
}
