% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabg.R
\name{cabg}
\alias{cabg}
\title{CABG}
\usage{
cabg(output_folder, anchor_date_table = NULL, before = NULL, after = NULL)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/cabg.csv
}
\description{
CABG
}
\details{
At least 1 CPT code:

33510, 33511, 33512, 33513, 33514, 33516, 33517, 33518, 33519, 33521, 33522, 33523,  33533, 33534, 33535, 33536
}
