% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cad.R, R/cad_alt.R
\name{cad}
\alias{cad}
\title{CAD}
\usage{
cad(output_folder, anchor_date_table = NULL, before = NULL, after = NULL)

cad(output_folder, anchor_date_table = NULL, before = NULL, after = NULL)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/cad.csv

output_folder/cad.csv
}
\description{
CAD

Alternate CAD Definition
}
\details{
At least 2 ICD codes and 1 CPT code.

ICD9: "410","410.\%","411","411.\%","412","412.\%","413","413.\%","414","414.\%","V45.82"

ICD10: "I25.1\%"

CPT: "33534","33535","33536","33510","33511","
33512","33513","33514","33515","33516",
"33517","33518","33519","33520","33521","33522",
"33523","92980","92981","92982","92984","92995","92996"

At least 2 ICD codes and 1 CPT code.

ICD9: 411.x, 412.x, 413.x,414.x,

ICD10: I20.\%, I25.\%

CPT: "33534","33535","33536","33510","33511","
33512","33513","33514","33515","33516",
"33517","33518","33519","33520","33521","33522",
"33523","92980","92981","92982","92984","92995","92996"
}
