% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcium_channel_blockers.R
\name{calcium_channel_blockers}
\alias{calcium_channel_blockers}
\title{Calcium channel blockers}
\usage{
calcium_channel_blockers(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
CSV saved to output_folder/calcium_channel_blockers.csv
}
\description{
Calcium channel blockers
}
\details{
Meds: "amlodipine","norvasc","diltiazem","cardizem","tiazac","felodipine","isradipine",
"nicardipine","cardenesr","cardene sr","cardene-sr","nifedipine","procardia","nisoldipine","sular",
"verapamil","calan","verelan","covera-hs","covera hs","coverahs","lotrel","caduet","bepridil","vascor",
"clevidipine","cleviprex","isoptin","plendil","lacidipine","caldine","lacimen","lacipil","midotens","motens",
"lercanidipine","lercadip","zanadip","dynacirc","adalat","nifediac","nifedical","nimodipine","nimotop",
"prestalia","teczem","lexxel","carmen ace","coripren","tarka","azor","tribenzor","twynsta","exforge","exforge hct"
}
