% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancer.R
\name{cancer}
\alias{cancer}
\title{Cancer}
\usage{
cancer(output_folder, anchor_date_table = NULL, before = NULL, after = NULL)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/cancer.csv
}
\description{
Cancer
}
\details{
At least 1 ICD code

ICD9: "104", "104.\%", "105", "105.\%", "106", "106.\%", "107", "107.\%",
"108", "108.\%", "109", "109.\%", "110", "110.\%", "111", "111.\%", "112",
"112.\%", "113", "113.\%", "114", "114.\%", "115", "115.\%", "116", "116.\%",
"117", "117.\%", "118", "118.\%", "119", "119.\%", "120", "120.\%", "121", "121.\%",
"122", "122.\%", "123", "123.\%", "124", "124.\%", "125", "125.\%", "126", "126.\%",
"127", "127.\%", "128", "128.\%", "129", "129.\%", "130", "130.\%", "131", "131.\%", "132", "132.\%",
"133", "133.\%", "134", "134.\%", "135", "135.\%", "136", "136.\%", "137", "137.\%", "138", "138.\%",
"139", "139.\%", "140", "140.\%", "141", "141.\%", "142", "142.\%", "143", "143.\%", "144", "144.\%",
"145", "145.\%", "146", "146.\%", "147", "147.\%", "148", "148.\%", "149", "149.\%", "150", "150.\%",
"151", "151.\%", "152", "152.\%", "153", "153.\%", "154", "154.\%", "155", "155.\%", "156", "156.\%",
"157", "157.\%", "158", "158.\%", "159", "159.\%", "160", "160.\%", "161", "161.\%", "162", "162.\%",
"163", "163.\%", "164", "164.\%", "165", "165.\%", "166", "166.\%", "167", "167.\%", "168", "168.\%",
"169", "169.\%", "170", "170.\%", "171", "171.\%", "172", "172.\%", "173", "173.\%", "174", "174.\%",
"175", "175.\%", "176", "176.\%", "177", "177.\%", "178", "178.\%", "179", "179.\%", "180", "180.\%",
"181", "181.\%", "182", "182.\%", "183", "183.\%", "184", "184.\%", "185", "185.\%", "186", "186.\%",
"187", "187.\%", "188", "188.\%", "189", "189.\%", "190", "190.\%", "191", "191.\%", "192", "192.\%",
"193", "193.\%", "194", "194.\%", "195", "195.\%", "196", "196.\%", "197", "197.\%", "198", "198.\%",
"199", "199.\%", "200", "200.\%", "201", "201.\%", "202", "202.\%", "203",
"203.\%", "204", "204.\%", "205", "205.\%", "206", "206.\%", "207", "207.\%", "208", "208.\%", "209", "209.\%"

ICD10: "C00", "C00.\%", "C01", "C01.\%", "C02", "C02.\%", "C03", "C03.\%", "C04", "C04.\%", "C05",
"C05.\%", "C06", "C06.\%", "C07", "C07.\%", "C08", "C08.\%", "C09", "C09.\%", "C10", "C10.\%", "C11",
"C11.\%", "C12", "C12.\%", "C13", "C13.\%", "C14", "C14.\%", "C15", "C15.\%", "C16", "C16.\%", "C17",
"C17.\%", "C18", "C18.\%", "C19", "C19.\%", "C20", "C20.\%", "C21", "C21.\%", "C22", "C22.\%", "C23",
"C23.\%", "C24", "C24.\%", "C25", "C25.\%", "C26", "C26.\%", "C27", "C27.\%", "C28", "C28.\%", "C29",
"C29.\%", "C30", "C30.\%", "C31", "C31.\%", "C32", "C32.\%", "C33", "C33.\%", "C34", "C34.\%", "C35",
"C35.\%", "C36", "C36.\%", "C37", "C37.\%", "C38", "C38.\%", "C39", "C39.\%", "C40", "C40.\%", "C41",
"C41.\%", "C42", "C42.\%", "C43", "C43.\%", "C44", "C44.\%", "C45", "C45.\%", "C46", "C46.\%", "C47",
"C47.\%", "C48", "C48.\%", "C49", "C49.\%", "C50", "C50.\%", "C51", "C51.\%", "C52", "C52.\%", "C53",
"C53.\%", "C54", "C54.\%", "C55", "C55.\%", "C56", "C56.\%", "C57", "C57.\%", "C58", "C58.\%", "C59",
"C59.\%", "C60", "C60.\%", "C61", "C61.\%", "C62", "C62.\%", "C63", "C63.\%", "C64", "C64.\%", "C65",
"C65.\%", "C66", "C66.\%", "C67", "C67.\%", "C68", "C68.\%", "C69", "C69.\%", "C70", "C70.\%", "C71",
"C71.\%", "C72", "C72.\%", "C73", "C73.\%", "C74", "C74.\%", "C75", "C75.\%", "C76", "C76.\%", "C77",
"C77.\%", "C78", "C78.\%", "C79", "C79.\%", "C80", "C80.\%", "C81", "C81.\%", "C82", "C82.\%", "C83",
"C83.\%", "C84", "C84.\%", "C85", "C85.\%", "C86", "C86.\%", "C87", "C87.\%", "C88", "C88.\%", "C89",
"C89.\%", "C90", "C90.\%", "C91", "C91.\%", "C92", "C92.\%", "C93", "C93.\%", "C94", "C94.\%", "C95",
"C95.\%", "C96", "C96.\%", "D00", "D00.\%", "D01", "D01.\%", "D02", "D02.\%", "D03", "D03.\%", "D04",
"D04.\%", "D05", "D05.\%", "D06", "D06.\%", "D07", "D07.\%", "D08", "D08.\%", "D09", "D09.\%", "D10",
"D10.\%", "D11", "D11.\%", "D12", "D12.\%", "D13", "D13.\%", "D14", "D14.\%", "D15", "D15.\%", "D16",
"D16.\%", "D17", "D17.\%", "D18", "D18.\%", "D19", "D19.\%", "D20", "D20.\%", "D21", "D21.\%", "D22",
"D22.\%", "D23", "D23.\%", "D24", "D24.\%", "D25", "D25.\%", "D26", "D26.\%", "D27", "D27.\%", "D28",
"D28.\%", "D29", "D29.\%", "D30", "D30.\%", "D31", "D31.\%", "D32", "D32.\%", "D33", "D33.\%", "D34",
"D34.\%", "D35", "D35.\%", "D36", "D36.\%", "D37", "D37.\%", "D38", "D38.\%", "D39", "D39.\%", "D40",
"D40.\%", "D41", "D41.\%", "D42", "D42.\%", "D43", "D43.\%", "D44", "D44.\%", "D45", "D45.\%", "D46",
"D46.\%", "D47", "D47.\%", "D48", "D48.\%", "D49", "D49.\%"
}
