% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronic_kidney_disease.R
\name{chronic_kidney_disease}
\alias{chronic_kidney_disease}
\title{chronic_kidney_disease}
\usage{
chronic_kidney_disease(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/chronic_kidney_disease.csv
}
\description{
chronic_kidney_disease
}
\details{
At least 1 ICD code:

ICD9: "585","585.\%"

ICD10: "N18","N18.\%"
}
