% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_bmi2.R
\name{closest_bmi2}
\alias{closest_bmi2}
\title{Closest BMI (Using only the BMI measure)}
\usage{
closest_bmi2(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/closest_bmi.csv
}
\description{
Closest BMI (Using only the BMI measure)
}
\details{
In closest_bmi algorithm, height and weight are used when BMI is missing. closest_bmi2 only uses the BMI measurement.
}
