% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_egfr.R
\name{closest_egfr}
\alias{closest_egfr}
\title{Closest eGFR}
\usage{
closest_egfr(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/closest_egfr.csv
}
\description{
Closest eGFR
}
\details{
Searches for
"Glomerular filtration rate/1.73 sq M.predicted \[Volume Rate/Area] in Serum, Plasma or Blood by Creatinine-based formula (MDRD)"
"Glomerular filtration rate/1.73 sq M.predicted among blacks \[Volume Rate/Area] in Serum, Plasma or Blood by Creatinine-based formula (MDRD)"
"Glomerular filtration rate/1.73 sq M.predicted among non-blacks \[Volume Rate/Area] in Serum, Plasma or Blood by Creatinine-based formula (MDRD)"
"Glomerular filtration rate/1.73 sq M.predicted \[Volume Rate/Area] in Serum, Plasma or Blood"
"Glomerular filtration rate/1.73 sq M.predicted \[Volume Rate/Area] in Serum, Plasma or Blood by Creatinine-based formula (CKD-EPI)"
"Glomerular filtration rate/1.73 sq M.predicted among non-blacks \[Volume Rate/Area] in Serum, Plasma or Blood by Creatinine-based formula (CKD-EPI)"
}
