% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_fasting_glucose.R
\name{closest_fasting_glucose}
\alias{closest_fasting_glucose}
\title{Closest fasting glucose}
\usage{
closest_fasting_glucose(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/closest_fasting_glucose.csv
}
\description{
Closest fasting glucose
}
\details{
Searches for

"Fasting glucose [Mass/volume] in Serum or Plasma"

"Fasting glucose [Moles/volume] in Blood"
}
