% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_glucose.R
\name{closest_glucose}
\alias{closest_glucose}
\title{Closest glucose}
\usage{
closest_glucose(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/closest_glucose.csv
}
\description{
Closest glucose
}
\details{
Searches for

"Glucose [Mass/volume] in Serum or Plasma"

"Glucose [Mass/volume] in Blood"
}
