% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_pro_bnp.R
\name{closest_pro_bnp}
\alias{closest_pro_bnp}
\title{Closest NT-PRO-BNP}
\usage{
closest_pro_bnp(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/closest_pro_bnp.csv
}
\description{
Closest NT-PRO-BNP
}
\details{
Searches for

"Natriuretic peptide.B prohormone N-Terminal [Mass/volume] in Serum or Plasma"
}
