% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copd_primary_inpatient_all.R
\name{copd_primary_inpatient_all}
\alias{copd_primary_inpatient_all}
\title{COPD primary inpatient all}
\usage{
copd_primary_inpatient_all(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/copd_primary_inpatient.csv
}
\description{
COPD primary inpatient all
}
\details{
Definition: Needs at least 1 inpatient ICD code:

ICD9: "491","491.0","491.1","491.2","491.20","491.21","491.22",
"491.8","491.9","492","492.0","492.8","496","496.0","493.21","493.22"

ICD10: "J44.\%","J43.\%","J42","J41.\%"
}
