% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_at_age_18.R
\name{date_at_age_18}
\alias{date_at_age_18}
\title{Date at Age 18}
\usage{
date_at_age_18(output_folder)
}
\arguments{
\item{output_folder}{the folder to write the output}
}
\value{
output_folder/date_at_age_18.csv
}
\description{
Date at Age 18
}
