% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcm_outcomes_afib.R
\name{dcm_outcomes_afib}
\alias{dcm_outcomes_afib}
\title{DCM OUTCOMES afib}
\usage{
dcm_outcomes_afib(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/dcm_outcomes_afib.csv
}
\description{
DCM OUTCOMES afib
}
\details{
At least 1 ICD code:

ICD9: '427.3','427.31'

ICD10: 'I48.0','I48.1','I48.11','I48.19','I48.2','I48.20','I48.21','I48.4','I48.9','I48.91'
}
