% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcm_outcomes_ventricular_arrhythmia.R
\name{dcm_outcomes_ventricular_arrhythmia}
\alias{dcm_outcomes_ventricular_arrhythmia}
\title{Afib Strict Definition}
\usage{
dcm_outcomes_ventricular_arrhythmia(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/afib_strict.csv
}
\description{
Afib Strict Definition
}
\details{
At least 2 ICD code:

ICD9: "427.3","427.31","427.32"

ICD10: "I48","I48.0","I48.1","I48.2","I48.3","I48.4","I48.9","I48.91","I48.92"

AND No pacemaker defined by at least 1 CPT code:
"0387T","0389T","0390T","0391T","33207","33208","33210","33211","33212","33213","33214","33221","33222","33226","33227",
"33228","33229","33233","33234","33235","33236","33237","33238","33274","33275",
"71090","93279","93280","93281","93286","93288","93293","93294","93731","93732","93733","93734","93735","93736"
}
