% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{demographics}
\alias{demographics}
\title{Demographics}
\usage{
demographics(output_folder)
}
\arguments{
\item{output_folder}{the folder to write the output}
}
\value{
output_folder/demographics.csv
}
\description{
Demographics
}
