% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics_robust.R
\name{demographics_robust}
\alias{demographics_robust}
\title{Demographics Robust}
\usage{
demographics_robust(output_folder)
}
\arguments{
\item{output_folder}{the folder to write the output}
}
\value{
output_folder/demographics_robust.csv
}
\description{
Demographics Robust
}
