% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{device}
\alias{device}
\title{Device information with syncing data}
\usage{
device(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL,
  cohort = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/device.csv
}
\description{
Device information with syncing data
}
\details{
Average and standard deviation of heart rate at each day
}
