% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diabetes.R
\name{diabetes}
\alias{diabetes}
\title{Diabetes}
\usage{
diabetes(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL,
  suffix = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}

\item{suffix}{optional string appended to the end of every output column name except person_id.}
}
\value{
output_folder/diabetes.csv
}
\description{
Diabetes
}
\details{
Based on the definition found in AOU phenotype library. Uses ICD9/10, meds, labs.
}
