% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diabetic_hhs.R
\name{diabetic_hhs}
\alias{diabetic_hhs}
\title{Diabetic hyperglycemic hyperosmolar syndrome (HHS)}
\usage{
diabetic_hhs(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/diabetic_hhs.csv
}
\description{
Diabetic hyperglycemic hyperosmolar syndrome (HHS)
}
\details{
At least 1 ICD code:

ICD9: 250.20, 250.21, 250.22, 250.23

ICD10:  E11.00, E11.01, E10.00, E10.01
}
