% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diabetic_ketoacidosis.R
\name{diabetic_ketoacidosis}
\alias{diabetic_ketoacidosis}
\title{Diabetic ketoacidosis}
\usage{
diabetic_ketoacidosis(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/diabetic_ketoacidosis.csv
}
\description{
Diabetic ketoacidosis
}
\details{
At least 1 ICD code:

ICD9: 250.10, 250.11, 250.12, 250.13

ICD10: E10.10, E10.11, E11.10, E11.11
}
