% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_bucket.R
\name{.read_from_bucket}
\alias{.read_from_bucket}
\title{Internal function used to read dataset from workspace bucket and local file system}
\usage{
.read_from_bucket(output_folder, name)
}
\arguments{
\item{output_folder}{the folder to read from}

\item{name}{the name of the file to read (without extension)}

\item{dat}{the data to read}
}
\description{
Internal function used to read dataset from workspace bucket and local file system
}
