% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyslipidemia.R
\name{dyslipidemia}
\alias{dyslipidemia}
\title{Dyslipidemia}
\usage{
dyslipidemia(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/dyslipidemia.csv
}
\description{
Dyslipidemia
}
\details{
Meets at least one condition:

If male and HDL < 40

If female and HDL < 50

LDL > 160

Triglycerides > 150

Cholesterol > 240
}
