% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employment.R
\name{employment}
\alias{employment}
\title{Employment}
\usage{
employment(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/employment.csv
}
\description{
Employment
}
