% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_light_sleep.R
\name{first_light_sleep}
\alias{first_light_sleep}
\title{First light sleep datetime}
\usage{
first_light_sleep(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/first_light_sleep.csv
}
\description{
First light sleep datetime
}
\details{
First light sleep level, date and time
}
