% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender_identity.R
\name{gender_identity}
\alias{gender_identity}
\title{Gender identity}
\usage{
gender_identity(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/gender_identity.csv
}
\description{
Gender identity
}
