% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gestational_diabetes.R
\name{gestational_diabetes}
\alias{gestational_diabetes}
\title{Gestational diabetes}
\usage{
gestational_diabetes(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/gestational_diabetes.csv
}
\description{
Gestational diabetes
}
\details{
At least 1 ICD code

ICD9: 648.00-648.04, 648.80-648.84

ICD10:  O24.410, O24.414, O24.419, O24.420, O24.424, O24.429, O24.430, O24.434, 24.439, O24.4. O24.41, O24.42, O24.43, O24.415, O24.435
}
