% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_insurance.R
\name{health_insurance}
\alias{health_insurance}
\title{Health insurance}
\usage{
health_insurance(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/health_insurance.csv
}
\description{
Health insurance
}
