% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heart_failure_specific.R
\name{heart_failure_specific}
\alias{heart_failure_specific}
\title{Heart failure, specific definition}
\usage{
heart_failure_specific(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/heart_failure.csv
}
\description{
Heart failure, specific definition
}
\details{
Removes codes that could be associated with cardiomyopathy.
At least 1 inpatient or 2 outpatient ICD codes:

ICD9: "428","428.\%",

ICD10: "I50","I50.\%"
}
