% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heart_rate_summary.R
\name{heart_rate_summary}
\alias{heart_rate_summary}
\title{Heart Rate Summary}
\usage{
heart_rate_summary(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/heart_rate_summary.csv
}
\description{
Heart Rate Summary
}
\details{
Heart rate summary includes for each person_id and date:
heart_rate_summary_zone_name
heart_rate_summary_min_heart_rate
heart_rate_summary_max_heart_rate
heart_rate_summary_minute_in_zone
}
