% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hermes_three.R
\name{hermes_three}
\alias{hermes_three}
\title{HERMES 3}
\usage{
hermes_three(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/hermes_three.csv
}
\description{
HERMES 3
}
\details{
At least 1 ICD code:

ICD9: "410","410.\%","411","411.\%"

ICD10: "I21","I21.\%","I22","I22.\%","I23","I23.\%","I24","I24.\%","I25","I25.\%"

Snomed: "85898001","52029003","471890009","195021004","53043001",
"204981000119101","30496006","74249003","111285003","1149109001","16253001",
"72972005","713523008","880042006","1234750000","12563008","766883006",
"471841009","62377009","16253001","153941000119100","417996009","703275009",
"703272007","703274008","443254009","153931000119109","443253003","441481004",
"120871000119108","120861000119102","120851000119104","1064941000000100"
}
