% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_min_heart_rate.R
\name{hourly_min_heart_rate}
\alias{hourly_min_heart_rate}
\title{Hourly minimum heart rate at each day}
\usage{
hourly_min_heart_rate(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL,
  cohort = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}

\item{cohort}{a vector of ids to limit the query}
}
\value{
output_folder/hourly_min_heart_rate.csv
}
\description{
Hourly minimum heart rate at each day
}
\details{
Minimum heart rate at each day and each hour
}
