% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_sdann.R
\name{hourly_sdann}
\alias{hourly_sdann}
\title{Hourly SDANN: Standard deviation of the average NN intervals}
\usage{
hourly_sdann(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/sdann.csv
}
\description{
Hourly SDANN: Standard deviation of the average NN intervals
}
\details{
Heart rate minute level data is extracted and five-minute averages are calculated for consecutive five-minute intervals.
The average HR is used to calculate the average RR duration of each five-minute interval:
Average RR = 6000 / mean(HR)
Subsequently, the standard deviation of all the five-minute RR intervals is calculated, yielding the SDANN value (in ms).
}
