% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibd.R
\name{ibd}
\alias{ibd}
\title{Inflammatory bowel disease (IBD)}
\usage{
ibd(output_folder, anchor_date_table = NULL, before = NULL, after = NULL)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/ibd.csv
}
\description{
Inflammatory bowel disease (IBD)
}
\details{
At least 1 ICD code:

ICD9: 555.0, 555.1, 555.2, 555.9, 556.0, 556.1, 556.2, 556.9,

ICD10: K50.0, K50.1, K50.8, K50.9, K51.0, K51.1, K51.2, K51.9
}
