% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insomnia.R
\name{insomnia}
\alias{insomnia}
\title{Insomnia}
\usage{
insomnia(output_folder, anchor_date_table = NULL, before = NULL, after = NULL)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/insomnia.csv
}
\description{
Insomnia
}
\details{
At least 1 ICD code:

ICD9: "307.41","327.01","327.02","327.09","780.52","307.42","327.0","327.00"

ICD10: "F51.04","G47.00","G47.0","G47.09","F51.05","G47.01","F51.02","F51.03","F51.01","F51.09"
}
