% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interstitial_lung_disease.R
\name{interstitial_lung_disease}
\alias{interstitial_lung_disease}
\title{Interstitial lung disease}
\usage{
interstitial_lung_disease(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/interstitial_lung_disease.csv
}
\description{
Interstitial lung disease
}
\details{
At least 1 ICD code:

ICD9: "515.\%","516.\%","135","501","508.1","518.89"
ICD10: "J84.\%","D86.0","J61","J70.1","J98.4"
}
