% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_algorithms_and_combine.R
\name{left_join_outputs}
\alias{left_join_outputs}
\title{Left join multiple tables onto a main table by person_id}
\usage{
left_join_outputs(main_df, ..., key = "person_id")
}
\arguments{
\item{main_df}{main data.frame/data.table defining the row universe}

\item{...}{additional tables to left join onto main_df}

\item{key}{join key (default "person_id")}
}
\value{
data.table
}
\description{
Left join multiple tables onto a main table by person_id
}
