% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_inpatient_outpatient.R
\name{mi_inpatient_outpatient}
\alias{mi_inpatient_outpatient}
\title{MI Inpatient/Outpatient}
\usage{
mi_inpatient_outpatient(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/mi_inpatient_outpatient.csv
}
\description{
MI Inpatient/Outpatient
}
\details{
At least 1 inpatient or 2 outpatient ICD codes:

ICD9: "410","410.\%","411","411.\%"

ICD10: "I21","I21.\%","I22","I22.\%","I23","I23.\%","I24","I24.\%","I25","I25.\%"
}
