% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mineralocorticoid_antagonists.R
\name{mineralocorticoid_antagonists}
\alias{mineralocorticoid_antagonists}
\title{Mineralocorticoid Antagonists}
\usage{
mineralocorticoid_antagonists(
  output_path,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}

\item{output_folder}{the folder to write the output}
}
\value{
output_folder/mineralocorticoid_antagonists.csv
}
\description{
Mineralocorticoid Antagonists
}
\details{
Meds: "spironolactone","aldactone","eplerenone","inspra"
}
