% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_recent_bmi.R
\name{most_recent_bmi}
\alias{most_recent_bmi}
\title{Most Recent BMI}
\usage{
most_recent_bmi(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL,
  suffix = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{optional data.frame containing columns: person_id, anchor_date.}

\item{before}{an integer >= 0}

\item{after}{an integer >= 0}

\item{suffix}{optional string appended to the end of every output column name except person_id.}
}
\value{
output_folder/most_recent_bmi.csv
}
\description{
Most Recent BMI
}
