% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_recent_hgba1c.R
\name{most_recent_hgba1c}
\alias{most_recent_hgba1c}
\title{Most Recent HgbA1c}
\usage{
most_recent_hgba1c(output_folder)
}
\arguments{
\item{output_folder}{the folder to write the output}
}
\value{
output_folder/most_recent_hgba1c.csv
}
\description{
Most Recent HgbA1c
}
\details{
Searches for

"Hemoglobin A1c/Hemoglobin.total in Blood by Electrophoresis"

"Hemoglobin A1c/Hemoglobin.total in Blood by calculation"

"Hemoglobin A1c/Hemoglobin.total in Blood by IFCC protocol"

"Hemoglobin A1c/Hemoglobin.total in Blood"

"Hemoglobin A1c/Hemoglobin.total in Blood by HPLC"
}