% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_query.R
\name{multi_query}
\alias{multi_query}
\title{All Multiple Queries}
\usage{
multi_query(
  query_list,
  output_folder,
  load_queries = FALSE,
  override = FALSE,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{query_list}{a list of names of query functions to run}

\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}

\item{either}{to override previous results}
}
\value{
query results saved as output_folder/query_list[[i]].csv. If load_queries is set to true each query result is loaded to a named list
}
\description{
All Multiple Queries
}
\details{
Runs multiple queries given by a list of function names. Name in the function needs to be exactly same with query functions in phenotyper2 package.
}
