% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narcolepsy.R
\name{narcolepsy}
\alias{narcolepsy}
\title{Narcolepsy}
\usage{
narcolepsy(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/narcolepsy.csv
}
\description{
Narcolepsy
}
\details{
At least 1 ICD code:

ICD-10:
G47.41: Narcolepsy with cataplexy
G47.411: Narcolepsy with cataplexy
G47.419: Narcolepsy without cataplexy
G47.42: Narcolepsy in conditions classified elsewhere
G47.421: Narcolepsy in conditions classified elsewhere with cataplexy
G47.429: Narcolepsy in conditions classified elsewhere without cataplexy

ICD-9:
347.0: This code specifically covers narcolepsy, a medical classification under "Other Disorders of the Central Nervous System" (340-349).

347.01: This code specifically covers narcolepsy with cataplexy
347.00: This code specifically covers narcolepsy without cataplexy
}
