% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obesity.R
\name{obesity}
\alias{obesity}
\title{Obesity}
\usage{
obesity(output_folder, anchor_date_table = NULL, before = NULL, after = NULL)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/obesity.csv
}
\description{
Obesity
}
\details{
At least 1 ICD code:

ICD9: "278.0\%","278.1\%","V85.3\%","V85.4\%"

ICD10: "E66","E66.\%"
}
