% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overall_health_survey_promis_mental.R
\name{overall_health_survey_promis_mental}
\alias{overall_health_survey_promis_mental}
\title{Overall Health Survey: PROMIS Mental Health}
\usage{
overall_health_survey_promis_mental(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.
PROMIS Mental Health

The next two-item scale measures a person's self-reported mental health by evaluating their general mental wellbeing and their social satisfaction:
\itemize{
\item \strong{1585729}: How would you rate your mental health?
\item \strong{1585735}: How would you rate your satisfaction with your social activities and relationships?
}

Ref: http://www.healthmeasures.net/images/PROMIS/manuals/PROMIS_Global_Scoring_Manual.pdf}
}
\value{
output_folder/overal_health_survey_promis_mental.csv
}
\description{
Overall Health Survey: PROMIS Mental Health
}
