% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overall_health_survey_promis_phys.R
\name{overall_health_survey_promis_phys}
\alias{overall_health_survey_promis_phys}
\title{Overall Health Survey: PROMIS Physical Health}
\usage{
overall_health_survey_promis_phys(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.
This first scale is a two-item measure of one's personal rating of their physical health, and how well a participant is able to carry out everyday activities.

\if{html}{\out{<div class="sourceCode">}}\preformatted{ This scale is comprised of 2 question concept IDs, listed below. To get only the information relevant to PROMIS Phycial Health,
 we can filter the dataset to include only those question concept IDs

- **1585723**: How would you rate your physical health?
- **1585741**: To what extent are you able to carry out your everyday physical activities?

Ref: http://www.healthmeasures.net/images/PROMIS/manuals/PROMIS_Global_Scoring_Manual.pdf
}\if{html}{\out{</div>}}

For this scale, we will rate physical health on a scale of 1-5, with higher scores corresponding to better physical health.

For general physical health, we rate as follows:
\itemize{
\item 1: Poor
\item 2: Fair
\item 3: Good
\item 4: Very Good
\item 5: Excellent
}

For everyday activities, we rate as follows:
\itemize{
\item 1: Not at All
\item 2: A Little
\item 3: Moderately
\item 4: Mostly
\item 5: Completely
}

The summed score is then coverted to a t-score according to the PROMIS scoring manual}
}
\value{
output_folder/overal_health_survey_promis_phys.csv
}
\description{
Overall Health Survey: PROMIS Physical Health
}
