% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_inpatient_outpatient.R
\name{pad_inpatient_outpatient}
\alias{pad_inpatient_outpatient}
\title{PAD Inpatient/Outpatient}
\usage{
pad_inpatient_outpatient(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/pad_inpatient_outpatient.csv
}
\description{
PAD Inpatient/Outpatient
}
\details{
At least 1 inpatient or 2 outpatient ICD codes:

ICD9: "443.9","443.8"

ICD10: "I73.8","I73.9","I70.20","I70.21","I70.22"
}
