% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnes.R
\name{pnes}
\alias{pnes}
\title{Psychogenic Nonepileptic Seizures (PNES)}
\usage{
pnes(output_folder, anchor_date_table = NULL, before = NULL, after = NULL)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/pnes.csv
}
\description{
Psychogenic Nonepileptic Seizures (PNES)
}
\details{
At least 1 ICD code:

ICD9: 300.11, 300.12, 300.13

ICD10: F44.4, F44.5, F44.8, F44.9
}
