% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pneumonia.R
\name{pneumonia}
\alias{pneumonia}
\title{Pneumonia}
\usage{
pneumonia(output_folder, anchor_date_table = NULL, before = NULL, after = NULL)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/asthma.csv
}
\description{
Pneumonia
}
\details{
At least 1 ICD code:

ICD9: "480","480.\%","481","481.\%","482","482.\%","483","483.\%","484","484.\%","485","485.\%","486","486.\%","487","487.\%","488","488.\%"

ICD10: "J18.\%"
}
