% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pneumonia_primary_inpatient.R
\name{pneumonia_primary_inpatient}
\alias{pneumonia_primary_inpatient}
\title{Pneumonia primary inpatient}
\usage{
pneumonia_primary_inpatient(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/pneumonia_primary_inpatient.csv
}
\description{
Pneumonia primary inpatient
}
\details{
At least 1 ICD code:

ICD9: "480","480.\%","481","481.\%","482","482.\%","483","483.\%","484","484.\%","485","485.\%","486","486.\%","487.0","488.11","488.81"

ICD10:"J09.X1", "J09.X2", "J10.00", "J10.01", "J10.08", "J10.1", "J11.0", "J11.08", "J11.1", "J12.0", "J12.1", "J12.2", "J12.3",
"J12.81", "J12.89", "J12.9", "J13", "J13.\%", "J14", "J14.\%", "J15.3", "J15.4", "J15.7", "J15.9", "J16.0", "J16.8", "J18.0", "J18.1", "J18.8", "J18.9"
}
