% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pregnancy_hypertensive_disorders.R
\name{pregnancy_hypertensive_disorders}
\alias{pregnancy_hypertensive_disorders}
\title{Pregnancy hypertensive disorders}
\usage{
pregnancy_hypertensive_disorders(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/pregnancy_hypertensive_disorders.csv
}
\description{
Pregnancy hypertensive disorders
}
\details{
At least 1 ICD code

ICD9: 64240, 64241, 64242, 64243, 64244, 64250, 64251, 64252, 64253, 64254, 64260, 64261, 64262, 64263, 64264, 64270, 64271, 64272, 64273, 64274

ICD10:  O13, O14, O14.00, O14.02, O14.03, O14.04, O14.05, O14.10, O14.12, O14.13, O14.90, O14.92, O14.93, O14.14, O14.15, O14.94, O14.95, O14.1, O15.00, O15.02, O15.03, O15.1, O15.9, O14.20, O14.22, O14.23, O14.24, O14.25, O11, O11.1, O11.2, O11.3, O11.4, O11.5, O11.9
}
