% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_consent_date.R
\name{primary_consent_date}
\alias{primary_consent_date}
\title{Primary consent date}
\usage{
primary_consent_date(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/primary_consent_date.csv
}
\description{
Primary consent date
}
\details{
The date someone enrolled into the All of Us Research Program as a data contributing participant
}
