% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_query.R
\name{run_query}
\alias{run_query}
\title{Run Query}
\usage{
run_query(
  query_name,
  output_folder = "datasets",
  bucket = NULL,
  load_query = FALSE,
  override = FALSE,
  ...
)
}
\arguments{
\item{query_name}{a string of query name to run}

\item{output_folder}{the folder to write the output and check for the cache}

\item{bucket}{the bucket to check for cached query}

\item{load_query}{either to load the result}

\item{override}{either to override previous results}

\item{...}{optional parameter to be passed to the query}
}
\value{
query result saved as output_folder/query_name.csv. If load_query is set to TRUE, query result is also returned.
}
\description{
Run Query
}
\details{
Runs a query after checking the cache in the output_folder.
Name in the function needs to be exactly same with query functions in aou.phenotyper2 package.
}
