% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdoh_survey_discrimination.R
\name{sdoh_survey_discrimination}
\alias{sdoh_survey_discrimination}
\title{SDOH Survey: Discrimination}
\usage{
sdoh_survey_discrimination(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/sdoh_survey_discrimination.csv
}
\description{
SDOH Survey: Discrimination
}
\details{
The next 9 items summarize survey participants' experiences with discrimination in everyday life:

40192380: How often do people act as if they are afraid of you?
40192395: How often do people act as if they think you are dishonest?
40192416: How often do you receive poorer service than other people at restaurants or stores?
40192451: How often are you threatened or harassed?
40192466: How often are you treated with less courtesy than other people?
40192489: How often are you treated with less respect than other people?
40192490: How often do people act as if they think you are not smart?
40192496: How often do people act as if they're better than you are?
40192519: How often are you called names or insulted?

Refs:
Williams, D.R., Yu, Y., Jackson, J.S., and Anderson, N.B. “Racial Differences in Physical and Mental Health: Socioeconomic Status, Stress, and Discrimination.” Journal of Health Psychology. 1997; 2(3):335-351.
Taylor T.R., Kamarck T.W., Shiffman S. “Validation of the Detroit area study discrimination scale in a community sample of older African American adults: the Pittsburgh healthy heart project.” International Journal of Behavioral Medicine. 2004; 11:88–94.
}
