% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdoh_survey_food_insecurity.R
\name{sdoh_survey_food_insecurity}
\alias{sdoh_survey_food_insecurity}
\title{SDOH Survey: Food insecurity}
\usage{
sdoh_survey_food_insecurity(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/sdoh_survey_food_insecurity.csv
}
\description{
SDOH Survey: Food insecurity
}
\details{
The following 2 questions measure whether or not a participant's household was unable to consistently have food present at home when needed:

40192426: Were you worried whether the food you had bought just didn't last and you didn't have money to get more?
40192517: Were you worried whether your food would run out before you got money to buy more?
Refs: Erin R. Hager, Anna M. Quigg, Maureen M. Black, Sharon M. Coleman, Timothy Heeren, Ruth Rose-Jacobs, John T. Cook, Stephanie A. Ettinger de Cuba, Patrick H. Casey, Mariana Chilton, Diana B. Cutts, Alan F. Meyers, Deborah A. Frank; Development and Validity of a 2-Item Screen to Identify Families at Risk for Food Insecurity. Pediatrics July 2010; 126 (1): e26–e32. 10.1542/peds.2009-3146

Radandt NE, Corbridge T, Johnson DB, Kim AS, Scott JM, Coldwell SE. Validation of a Two-Item Food Security Screening Tool in a Dental Setting. J Dent Child (Chic). 2018 Sep 15;85(3):114-119. PMID: 30869587; PMCID: PMC6419517.#'
}
