% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdoh_survey_housing_instability.R
\name{sdoh_survey_housing_instability}
\alias{sdoh_survey_housing_instability}
\title{SDOH Survey: Housing instability}
\usage{
sdoh_survey_housing_instability(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/sdoh_survey_housing_instability.csv
}
\description{
SDOH Survey: Housing instability
}
\details{
This 1-item scale examines how often a survey respondant has moved in the past year:

40192441: Number of moves in past 12 months
Ref: https://www.aamc.org/media/25736/download
}
