% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdoh_survey_neighborhood_disorder.R
\name{sdoh_survey_neighborhood_disorder}
\alias{sdoh_survey_neighborhood_disorder}
\title{SDOH Survey: Neighborhood disorder}
\usage{
sdoh_survey_neighborhood_disorder(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{The folder to write the output.}

\item{anchor_date_table}{A data.frame containing two columns:
person_id and anchor_date. A time window can be defined around the anchor
date using the \code{before} and \code{after} arguments.}

\item{before}{An integer greater than or equal to 0. Dates prior to
anchor_date - before will be excluded.}

\item{after}{An integer greater than or equal to 0. Dates after
anchor_date + after will be excluded.}
}
\value{
output_folder/sdoh_survey_neighborhood_disorder.csv
}
\description{
SDOH Survey: Neighborhood disorder
}
\details{
The following 13 questions measure perceived physical and social
disorder in a participant's neighborhood:

40192420: There is a lot of graffiti in your neighborhood
40192522: Your neighborhood is noisy
40192412: Vandalism is common in your neighborhood
40192469: There are a lot of abandoned buildings in your neighborhood
40192456: Your neighborhood is clean
40192386: People in your neighborhood take good care of their houses and apartments
40192500: There are too many people hanging around on the streets near your home
40192493: There is a lot of crime in your neighborhood
40192457: There is too much drug use in your neighborhood
40192476: There is too much alcohol use in your neighborhood
40192404: You are always having trouble with your neighbors
40192400: People in your neighborhood watch out for each other
40192384: Your neighborhood is safe

The four positively worded items concerning cleanliness, property care,
neighbors watching out for one another, and neighborhood safety are
reverse-coded. The final score is the mean of all 13 items and ranges from
1 to 4. Higher scores indicate greater perceived neighborhood disorder.
Participants must have valid responses to all 13 items to receive a score.
}
