% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdoh_survey_loneliness.R
\name{sdoh_survey_panes_loneliness}
\alias{sdoh_survey_panes_loneliness}
\title{SDOH Survey: Loneliness, ULS-8}
\usage{
sdoh_survey_panes_loneliness(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/sdoh_survey_loneliness.csv
}
\description{
SDOH Survey: Loneliness, ULS-8
}
\details{
The following 8 questions comprise the loneliness scale, which evaluates the social satisfaction of survey participants:
40192390: How often do you feel that you are unhappy being so withdrawn?
40192397: How often do you feel that there is no one you can turn to?
40192398: How often do you feel left out?
40192494: How often do you feel that people are around you but not with you?
40192501: How often do you feel isolated from others?
40192504: How often do you feel that you are an outgoing person?
40192507: How often do you feel lack companionship?
40192516: How often do you fell that you can find companionship when you want it?

Ref: Hays RD, DiMatteo MR. A short-form measure of loneliness. J Pers Assess. 1987 Spring;51(1):69-81. doi: 10.1207/s15327752jpa5101_6. PMID: 3572711
}
