% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdoh_survey_panes_walk_bike.R
\name{sdoh_survey_panes_walk_bike}
\alias{sdoh_survey_panes_walk_bike}
\title{SDOH Survey: PANES - Walking and Bicycling}
\usage{
sdoh_survey_panes_walk_bike(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/sdoh_survey_panes_walk_bike.csv
}
\description{
SDOH Survey: PANES - Walking and Bicycling
}
\details{
40192410: My neighborhood has several free or low-cost recreation facilities
40192431: There are facilities to bicycle in or near my neighborhood
40192436: Many shops, stores, markets or other places to buy things I need are within easy walking distance of my home
40192437: There are sidewalks on most of the streets in my neighborhood
40192440: It is within a 10-15 minute walk to a transit stop
}
