% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdoh_survey_perceived_stress.R
\name{sdoh_survey_perceived_stress}
\alias{sdoh_survey_perceived_stress}
\title{SDOH Survey: Perceived stress}
\usage{
sdoh_survey_perceived_stress(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/sdoh_survey_perceived_stress.csv
}
\description{
SDOH Survey: Perceived stress
}
\details{
Cohen Perceived Stress Scale, a metric which measures how well participants handled their thoughts and feelings related to stress over the previous month:
40192381: How often have you felt that you were unable to control the important things in your life?
40192396: How often have you been angered because of things that were outside of your control?
40192419: How often have you felt confident about your ability to handle your personal problems?
40192445: How often have you felt that you were on top of things?
40192449: How often have you been able to control irritations in your life?
40192452: How often have you been upset because of something that happened unexpectedly?
40192462: How often have you felt difficulties were piling up so high that you could not overcome them?
40192491: How often have you felt nervous and stressed?
40192506: How often have you found that you could not cope with all the things that you had to do?
40192525: How often have you felt that things were going your way?
Ref: Cohen S, Kamarck T, Mermelstein R. A global measure of perceived stress. J Health Soc Behav. 1983 Dec;24(4):385-96. PMID: 6668417
}
