% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdoh_survey_social_cohesion.R
\name{sdoh_survey_social_cohesion}
\alias{sdoh_survey_social_cohesion}
\title{SDOH Survey: Social cohesion}
\usage{
sdoh_survey_social_cohesion(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/sdoh_survey_social_cohesion.csv
}
\description{
SDOH Survey: Social cohesion
}
\details{
This 4-item scale evaluates the perceptions of neighborhood characteristics:

40192411: People in your neighborhood generally get along with each other
40192417: People in your neighborhood share the same values
40192463: People around here are willing to help their neighbor
40192499: People in your neighborhood can be trusted
Ref: Mahasin S. Mujahid, Ana V. Diez Roux, Jeffrey D. Morenoff, Trivellore Raghunathan, Assessing the Measurement Properties of Neighborhood Scales: From Psychometrics to Ecometrics, American Journal of Epidemiology, Volume 165, Issue 8, 15 April 2007, Pages 858–867, https://doi.org/10.1093/aje/kwm040
}
