% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdoh_survey_social_support.R
\name{sdoh_survey_social_support}
\alias{sdoh_survey_social_support}
\title{SDOH Survey: Social support}
\usage{
sdoh_survey_social_support(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/sdoh_survey_social_support.csv
}
\description{
SDOH Survey: Social support
}
\details{
The following eight questions evaluate the supportive relationships that survey participants have in their lives:

40192388: How often do you have someone to prepare your meals if you were unable to do it yourself?
40192399: How often do you have someone who understands your problems?
40192439: How often do you have someone to have a good time with?
40192442: How often do you have someone to help you if you were confined to bed?
40192446: How often do you have someone to love and make you feel wanted?
40192480: How often do you have someone to take you to the doctor if you need it?
40192511: How often do you have someone to help you with daily chores if you were sick?
40192528: How often do you have someone to turn to for suggestions about how to deal with a personal problem?
Ref: https://www.rand.org/health-care/surveys_tools/mos/social-support/survey-instrument.html
}
