% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdoh_survey_spiritual_exp.R
\name{sdoh_survey_spiritual_exp}
\alias{sdoh_survey_spiritual_exp}
\title{SDOH Survey: Daily spiritual experiences}
\usage{
sdoh_survey_spiritual_exp(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/sdoh_survey_spiritual_exp.csv
}
\description{
SDOH Survey: Daily spiritual experiences
}
\details{
This 6-item scale relates to the impact of religion and spirituality on the lives of survey participants:

40192401: How often do you feel deep inner peace or harmony?
40192415: How often do you feel that you are spiritually touched by the beauty of creation?
40192443: How often do you desire to be closer to or in union with God (or a higher power)?
40192471: How often do you feel God's (or a higher power's) love for you, directly or through others?
40192475: How often do you find strength and comfort in your religion?
40192498: How often do you feel God's (or a higher power's) presence?
Ref: Masters K.S. (2013) Brief Multidimensional Measure of Religiousness/Spirituality (BMMRS). In: Gellman M.D., Turner J.R. (eds) Encyclopedia of Behavioral Medicine. Springer, New York, NY. https://doi.org/10.1007/978-1-4419-1005-9_1577
}
