% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses.R
\name{ses}
\alias{ses}
\title{Socioeconomic status variables}
\usage{
ses(output_folder, anchor_date_table = NULL, before = NULL, after = NULL)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. The dataset will be merged with demographics.}

\item{before}{serves no function}

\item{after}{serves no function}
}
\value{
output_folder/ses.csv
}
\description{
Socioeconomic status variables
}
