% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sexual_orientation.R
\name{sexual_orientation}
\alias{sexual_orientation}
\title{Sexual orientation}
\usage{
sexual_orientation(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/sexual_orientation.csv
}
\description{
Sexual orientation
}
