% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sglt2_inhibitors.R
\name{sglt2_inhibitors}
\alias{sglt2_inhibitors}
\title{SGLT2 Inhibitors}
\usage{
sglt2_inhibitors(
  output_path,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}

\item{output_folder}{the folder to write the output}
}
\value{
output_folder/sglt2_inhibitors.csv
}
\description{
SGLT2 Inhibitors
}
\details{
Meds: "empagliflozin","jardiance","canagliflozin","invokana","dapagliflozin","farxiga","ertugliflozin","steglatro"
}
