% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep_level_deep.R
\name{sleep_level_deep}
\alias{sleep_level_deep}
\title{Deep sleep level data from Fitbit}
\usage{
sleep_level_deep(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/sleep_level_deep.csv
}
\description{
Deep sleep level data from Fitbit
}
\details{
Fitbit sleep data contains:
person_id
sleep_date
start_datetime
is_main_sleep
duration_in_min
}
