% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep_meds.R
\name{sleep_meds}
\alias{sleep_meds}
\title{Sleep meds}
\usage{
sleep_meds(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
CSV saved to output_folder/sleep_meds.csv
}
\description{
Sleep meds
}
\details{
Meds: benzodiazepines: temazepam,restoril,lorazepam,ativan,diazepam,valium
non-benzodiazepine_hypnotics: zolpidem,ambien,ambien cr,eszopiclone,lunesta,zaleplon,sonata
melatonin_receptor_agonists: suvorexant,belsomra,lemborexant,dayvigo
antidepressants_off_label: trazodone,desyrel,doxepin,silenor
antihistamines: diphenhydramine,benadryl,nytol,doxylamine,unisom
melatonin: melatonin
}
