% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking_current.R
\name{smoking_current}
\alias{smoking_current}
\title{Current smoking status}
\usage{
smoking_current(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/smoking_current.csv
}
\description{
Current smoking status
}
\details{
Data is from the AoU intake survey. Do you now smoke cigarettes every day, some days, or not at all?
}
