% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable_house_concern.R
\name{stable_house_concern}
\alias{stable_house_concern}
\title{Stable house concern}
\usage{
stable_house_concern(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/stable_house_concern.csv
}
\description{
In the past 6 months, have you been worried or concerned about NOT having a place to live?
}
