% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statins.R
\name{statins}
\alias{statins}
\title{Statins}
\usage{
statins(output_folder, anchor_date_table = NULL, before = NULL, after = NULL)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing three columns: person_id, anchor_date, statins_status. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/beta_blockers.csv
}
\description{
Statins
}
\details{
Meds: "atorvastatin","lipitor","simvastatin","zocor","
rosuvastatin","crestor","pravastatin","pravachol","fluvastatin",
"lescol","lovastatin","mevacor","vytorin","caduet"
}
