% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_count_average_5_minute_interval.R
\name{step_count_average_5_minute_interval}
\alias{step_count_average_5_minute_interval}
\title{Average step count over 5 minute intervals}
\usage{
step_count_average_5_minute_interval(output_folder)
}
\arguments{
\item{output_folder}{the folder to write the output}
}
\value{
output_folder/sdann.csv
}
\description{
Average step count over 5 minute intervals
}
\details{
Average step count over all available 5 minute intervals in a given day.
}
