% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stroke_inpatient_outpatient.R
\name{stroke_inpatient_outpatient}
\alias{stroke_inpatient_outpatient}
\title{Stroke Inpatient/Outpatient}
\usage{
stroke_inpatient_outpatient(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/stroke_inpatient_outpatient.csv
}
\description{
Stroke Inpatient/Outpatient
}
\details{
At least 1 inpatient or 2 outpatient ICD codes:

ICD9: "433.01","433.11","433.21","433.31","433.81","433.91","434.01","434.11","434.91","436.\%"

ICD10: "I63.\%","G46.3","G46.4"
}
