% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top10_steps.R
\name{top10_steps}
\alias{top10_steps}
\title{Top 10 Steps}
\usage{
top10_steps(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\value{
output_folder/top10_steps.csv
}
\description{
Top 10 Steps
}
