% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgs.R
\name{wgs}
\alias{wgs}
\title{Whole Genome Sequencing Flag}
\usage{
wgs(output_folder)
}
\arguments{
\item{output_folder}{the folder to write the output}
}
\value{
output_folder/wgs.csv
}
\description{
Whole Genome Sequencing Flag
}
\details{
wgs.csv will contain:
person_id
wgs_status
}
