% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayer_landmark.R
\name{bayer_landmark}
\alias{bayer_landmark}
\title{BMI and eGFR Two-Year Follow-up}
\usage{
bayer_landmark(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL,
  suffix = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{optional data.frame containing columns: person_id, anchor_date.}

\item{before}{an integer >= 0}

\item{after}{an integer >= 0}

\item{suffix}{optional string appended to the end of every output column name except person_id.}
}
\value{
output_folder/bayer_landmark.csv
}
\description{
BMI and eGFR Two-Year Follow-up
}
\details{
For each participant, defines:
t0 = date of first BMI,
t1 = t0 + 1 year,
t2 = t1 + 1 year.

Sets \code{bayer_landmark} to TRUE when participants have:
\itemize{
\item at least 1 BMI in (t0, t1]
\item at least 1 BMI in (t1, t2]
\item at least 1 eGFR in [t0, t2]
}
}
