% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_bucket.R
\name{.write_to_bucket}
\alias{.write_to_bucket}
\title{Internal function used to write dataset to workspace bucket and local file system}
\usage{
.write_to_bucket(dat, output_folder, name)
}
\description{
Internal function used to write dataset to workspace bucket and local file system
}
