% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfpef_icd.R
\name{hfpef_icd}
\alias{hfpef_icd}
\title{HFpEF using ICD codes}
\usage{
hfpef_icd(output_folder, anchor_date_table = NULL, before = NULL, after = NULL)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/hfpef.csv
}
\description{
HFpEF using ICD codes
}
\details{
At least 1 ICD code

Inclusion: 1 inpatient or 2 outpatient
ICD-9 Codes
•	428.30 Unspecified diastolic heart failure
•	428.31 Acute diastolic heart failure
•	428.32 Chronic diastolic heart failure
•	428.33 Acute on chronic diastolic heart failure
ICD-10 Codes
•	I50.30 Unspecified diastolic heart failure
•	I50.31 Acute diastolic heart failure
•	I50.32 Chronic diastolic heart failure
•	I50.33 Acute on chronic diastolic heart failure
Exclusion Criterion: HF reduced EF 1 inpatient or 2 outpatient
ICD-9 Codes
•	428.20 Unspecified systolic heart failure
•	428.21 Acute systolic heart failure
•	428.22 Chronic systolic heart failure
•	428.23 Acute on chronic systolic heart failure
ICD-10 Codes
•	I50.20 Unspecified systolic heart failure
•	I50.21 Acute systolic heart failure
•	I50.22 Chronic systolic heart failure
•	I50.23 Acute on chronic systolic heart failure
}
