% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_recent_bmi.R
\name{most_recent_bmi}
\alias{most_recent_bmi}
\title{Most Recent BMI}
\usage{
most_recent_bmi(output_folder)
}
\arguments{
\item{output_folder}{the folder to write the output}
}
\value{
output_folder/most_recent_bmi.csv
}
\description{
Most Recent BMI
}