% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_recent_egfr.R
\name{most_recent_egfr}
\alias{most_recent_egfr}
\title{Most Recent eGFR}
\usage{
most_recent_egfr(output_folder)
}
\arguments{
\item{output_folder}{the folder to write the output}
}
\value{
output_folder/most_recent_egfr.csv
}
\description{
Most Recent eGFR
}
\details{
Searches for

"Glomerular filtration rate/1.73 sq M.predicted [Volume Rate/Area] in Serum, Plasma or Blood by Creatinine-based formula (MDRD)"

"Glomerular filtration rate/1.73 sq M.predicted among blacks [Volume Rate/Area] in Serum, Plasma or Blood by Creatinine-based formula (MDRD)"

"Glomerular filtration rate/1.73 sq M.predicted among non-blacks [Volume Rate/Area] in Serum, Plasma or Blood by Creatinine-based formula (MDRD)"

"Glomerular filtration rate/1.73 sq M.predicted [Volume Rate/Area] in Serum, Plasma or Blood"

"Glomerular filtration rate/1.73 sq M.predicted [Volume Rate/Area] in Serum, Plasma or Blood by Creatinine-based formula (CKD-EPI)"

"Glomerular filtration rate/1.73 sq M.predicted among non-blacks [Volume Rate/Area] in Serum, Plasma or Blood by Creatinine-based formula (CKD-EPI)"
}