% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_query.R
\name{multi_query}
\alias{multi_query}
\title{Multiple Queries}
\usage{
multi_query(
  query_list,
  output_folder = "datasets",
  bucket = NULL,
  load_queries = FALSE,
  override = FALSE,
  merge_queries = FALSE,
  merge_on = "person_id",
  ...
)
}
\arguments{
\item{query_list}{a list of names of query functions to run}

\item{output_folder}{the folder to write the output}

\item{bucket}{the bucket to check for cached query}

\item{load_queries}{either to load results of the queries into the environment. If set true function will read each resulting query csv file and return the result. the names in list are the same as query_list}

\item{override}{either to override previous results}

\item{merge_queries}{either to merge queries. This uses a full merge to maintain all query information therefore might cause a cartesian product. If load_queries is set to false this parameter is ignored.}

\item{merge_on}{a column name to merge queries on. If load_queries or merge_queries is set to false this parameter is ignored.}

\item{...}{optional parameter to be passed to the query. If a query gets an unexpected parameter this may raise an error.}
}
\value{
query results saved as output_folder/query_list\[\[i]].csv.
If load_queries is set to true and merge_queries is false a named list with each query result is returned
If merge_queries is not TRUE the query results are merged on the given column name and the resulting data table is returned.
}
\description{
Multiple Queries
}
\details{
Runs multiple queries given by a list of function names.
Name in the function needs to be exactly same with query functions in aou.phenotyper2 package.
}
