% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad.R
\name{pad}
\alias{pad}
\title{Peripheral artery disease}
\usage{
pad(output_folder, anchor_date_table = NULL, before = NULL, after = NULL)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/pad.csv
}
\description{
Peripheral artery disease
}
\details{
At least 1 ICD code:

ICD9: 440.2, 440.20, 440.21, 440.22, 440.23, 440.24, 440.29, 443.9, 444.01, 444.09, 444.1, 444.21, 444.22, 444.81, 444.89, 444.9
ICD10 : I70.2xx, I73.9, I74.x
}
