% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_query.R
\name{run_query}
\alias{run_query}
\title{Run Query}
\usage{
run_query(
  query_name,
  output_folder,
  bucket = NULL,
  load_query = FALSE,
  override = FALSE,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{query_name}{a string of query name to run}

\item{output_folder}{the folder to write the output and check for the cache}

\item{bucket}{the bucket to check for cached query}

\item{load_query}{either to load the result}

\item{override}{either to override previous results}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
query result saved as output_folder/query_name.csv. If load_query is set to TRUE, query result is also returned.
}
\description{
Run Query
}
\details{
Runs a query after checking the cache in the output_folder.
Name in the function needs to be exactly same with query functions in aou.phenotyper2 package.
}
