% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdoh_survey_discrimination_medical.R
\name{sdoh_survey_discrimination_medical}
\alias{sdoh_survey_discrimination_medical}
\title{SDOH Survey: Discrimination in medical settings}
\usage{
sdoh_survey_discrimination_medical(
  output_folder,
  anchor_date_table = NULL,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{output_folder}{the folder to write the output}

\item{anchor_date_table}{a data.frame containing two columns: person_id, anchor_date. A time window can be defined around the anchor date using the \code{before} and \code{after} arguments.}

\item{before}{an integer greater than or equal to 0. Dates prior to anchor_date + before will be excluded.}

\item{after}{an integer greater than or equal to 0. Dates after anchor_date + after will be excluded.}
}
\value{
output_folder/sdoh_survey_discrimination_medical.csv
}
\description{
SDOH Survey: Discrimination in medical settings
}
\details{
7 items evaluate discrimination specifically in healthcare settings:

40192383: How often does a doctor or nurse act as if he or she is better than you?
40192394: How often do you feel like a doctor or nurse is not listening to what you were saying?
40192423: How often does a doctor or nurse act as if he or she is afraid of you?
40192425: How often are you treated with less respect than other people?
40192497: How often are you treated with less courtesy than other people?
40192505: How often does a doctor or nurse act as if he or she thinks you are not smart?
}
